Test the operator precedence
Because of originally using "data++", just add the address of data
pointing to. We want to add the value of the data pointing to.
Therefore, we use dereference like '*data++'.

Actually, according to spec of c language, the precedence of suffix
increment is higher than the one of dereference. It is still the
wrong way to add value of data pointing to.